+++ src/main/java/pro/sky/calcul/CalculatorService.java
@@ -9,5 +9,5 @@ public interface CalculatorService {
 
     int multiply(int num1, int num2);
 
-    int divide(int num1, int num2);
+    double divide(double num1, double num2);
 }
